Update .NET version and adjust PostCard border thickness
Updated .NET version in GitHub Actions workflow from 8.x.x to 9.x.x
in multiple job definitions within the build.yml file. This ensures
the project builds with the latest .NET environment.

Modified the BorderThickness property of the PostCard element in
Home.fs. The top border thickness was increased from 0 to 1, changing
the values from (1, 0, 0, 2) to (1, 1, 1, 2).